feat: 날짜별 점수 누적 기능 구현
Room @Upsert를 이용하여 현재 점수를 당일의 이전 점수에 누적하였습니다.

+++ app/src/main/java/com/example/quietnight/ui/QuietNightContract.kt
@@ -11,7 +11,7 @@ sealed class SnoreIntent {
 data class SnoreState(
     // Home State
     val todayScore: Int = 0,
-    val todaySnoreTime: Int = 0,
+    val todaySnoreTime: Long = 0,
     val todaySnoreMax: Int = 0,
     val positionStats: Map<String, Float> = emptyMap(),
